// frontend-next/lib/types/app-config.ts
/**
 * Response type from /api/app/about endpoint
 */
export interface AppConfig {
  production: boolean;
  version: string;
  demoStatus: boolean;
  allowSignup: boolean;
  allowPasswordLogin: boolean;
  defaultGroupSlug: string;
  defaultHouseholdSlug: string;
  enableOidc: boolean;
  oidcRedirect: boolean;
  oidcProviderName: string;
  enableOpenai: boolean;
  enableOpenaiImageServices: boolean;
}